Fix #10

+++ internal/unit/weight.go
@@ -3,5 +3,5 @@ package unit
 type Weight float64
 
 func (r Weight) Relatedness() Relatedness {
-	return Relatedness(1.0 / float64(r))
+	return Relatedness(1.0 / r)
 }
